Moved content related classes to R2APIContentManagment, fixed typos and issues.
