+++ R2API.Test/Tests/ReflectionTests.cs
@@ -1,8 +1,8 @@
+using R2API.Utils;
+using RoR2;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using R2API.Utils;
-using RoR2;
 using Xunit;
 
 namespace R2API.Test {
