+++ R2API/BuffAPI.cs
@@ -1,4 +1,5 @@
-﻿using R2API.Utils;
+﻿using R2API.ContentManagment;
+using R2API.Utils;
 using RoR2;
 using System;
 using System.Collections.ObjectModel;
