+++ R2API/EffectAPI.cs
@@ -1,3 +1,4 @@
+using R2API.ContentManagment;
 using R2API.Utils;
 using RoR2;
 using System;
