+++ R2API/EliteAPI.cs
@@ -1,4 +1,5 @@
 ﻿using MonoMod.Cil;
+using R2API.ContentManagment;
 using R2API.Utils;
 using RoR2;
 using System;
