+++ R2API/LoadoutAPI.cs
@@ -1,4 +1,5 @@
 using EntityStates;
+using R2API.ContentManagment;
 using R2API.Utils;
 using RoR2;
 using RoR2.Skills;
