+++ R2API/ProjectileAPI.cs
@@ -1,4 +1,5 @@
-﻿using R2API.Utils;
+﻿using R2API.ContentManagment;
+using R2API.Utils;
 using System;
 using System.Reflection;
 using UnityEngine;
