+++ R2API/SoundAPI.cs
@@ -1,6 +1,7 @@
 using BepInEx;
 using MonoMod.Cil;
 using MonoMod.RuntimeDetour;
+using R2API.ContentManagment;
 using R2API.MiscHelpers;
 using R2API.Utils;
 using RoR2;
